+++ amazonorders/__init__.py
@@ -1,3 +1,3 @@
 __copyright__ = "Copyright (c) 2024 Alex Laird"
 __license__ = "MIT"
-__version__ = "1.1.1"
+__version__ = "1.1.2"