+++ CHANGELOG.md
@@ -6,10 +6,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Unreleased](https://github.com/alexdlaird/amazon-orders/compare/1.1.1...HEAD)
 
+## [1.1.2](https://github.com/alexdlaird/amazon-orders/compare/1.1.1...1.1.2) - TBD
+
 ### Added
 
+- Raise `AmazonOrdersNotFoundError` when Order is not found.
+- Build improvements.
 - Documentation improvements.
-- Test improvements (Amazon no longer provider the `condition` field in many cases).
+- Test improvements (Amazon no longer provides the `condition` field in many cases).
 
 ### Changed